Token auth - login, permissions, and minor fixes

+++ base/views.py
@@ -1,13 +1,16 @@
 from rest_framework import status
-from rest_framework.decorators import api_view
+from rest_framework.decorators import api_view,permission_classes
 from rest_framework.response import Response
 from base.models import Note
 from base.serializers import NoteSerializer
+from rest_framework import generics
+from rest_framework.permissions import IsAuthenticatedOrReadOnly
 
 
 
 #View to create new note or list all existing notes
 @api_view(['GET', 'POST'])
+@permission_classes([IsAuthenticatedOrReadOnly])
 def note_list(request):
     
     #fetch all existing notes
@@ -20,13 +23,14 @@ def note_list(request):
     elif request.method == 'POST':
         serializer = NoteSerializer(data=request.data)
         if serializer.is_valid():
-            serializer.save()
+            serializer.save(created_by=request.user)
             return Response(serializer.data, status=status.HTTP_201_CREATED)
         return Response(serializer.errors, status=status.HTTP_400_BAD_REQUEST)
 
 
 #View to fetch, update or delete an existing note
 @api_view(['GET', 'PUT', 'DELETE'])
+@permission_classes([IsAuthenticatedOrReadOnly])
 def note_detail(request, pk):
     
     #check if note exists with id(pk)
@@ -35,19 +39,27 @@ def note_detail(request, pk):
     except Note.DoesNotExist:
         return Response(status=status.HTTP_404_NOT_FOUND)
     
-    #return note detail
+    #read note detail
     if request.method == 'GET':
         serializer = NoteSerializer(note)
         return Response(serializer.data)
 
     #update note
     elif request.method == 'PUT':
-        serializer = NoteSerializer(note, data=request.data)
-        if serializer.is_valid():
-            serializer.save()
-            return Response(serializer.data)
+        if request.user == note.created_by:
+            serializer = NoteSerializer(note, data=request.data)
+            if serializer.is_valid():
+                serializer.save()
+                return Response(serializer.data)
+        else:
+            return Response(status=status.HTTP_401_UNAUTHORIZED)
         return Response(serializer.errors, status=status.HTTP_400_BAD_REQUEST)
-    #delte note
+    #delete note
     elif request.method == 'DELETE':
-        note.delete()
-        return Response(status=status.HTTP_204_NO_CONTENT)
+        if request.user == note.created_by:
+            note.delete()
+            return Response(status=status.HTTP_204_NO_CONTENT)
+        else:
+            return Response(status=status.HTTP_401_UNAUTHORIZED,)
+        
+
